LPS-131457 This should be an int

+++ modules/apps/info/info-api/src/main/java/com/liferay/info/pagination/InfoPage.java
@@ -44,7 +44,7 @@ public class InfoPage<T> {
 		return _pageItems;
 	}
 
-	public long getTotalCount() {
+	public int getTotalCount() {
 		return _totalCountSupplier.get();
 	}
